src/pipeline-ledger/pyproject.toml
[tool.poetry]
name = "pipeline-ledger"
version = "0.1.0"
description = "Immutable append-only ledger for ORBIT AI Pipeline events"
authors = ["AdPorts AI Platform Team"]
python = "^3.12"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
esdbclient = "^1.1.0"
confluent-kafka = "^2.5.0"
asyncpg = "^0.30.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.35" }
alembic = "^1.13.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
httpx = "^0.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
structlog = "^24.4.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
